Initial changes

+++ src/Compute/Compute/Extension/Diagnostics/RemoveAzureRmVMDiagnosticsExtension.cs
@@ -57,6 +57,9 @@ namespace Microsoft.Azure.Commands.Compute
         [ResourceNameCompleter("Microsoft.Compute/virtualMachines/extensions", "ResourceGroupName", "VMName")]
         public string Name { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Returns immediately with status of request")]
+        public SwitchParameter NoWait { get; set; }
+
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -84,12 +87,24 @@ namespace Microsoft.Azure.Commands.Compute
                     }
                 }
 
-                var op = this.VirtualMachineExtensionClient.DeleteWithHttpMessagesAsync(
-                    this.ResourceGroupName,
-                    this.VMName,
-                    this.Name).GetAwaiter().GetResult();
-                var result = ComputeAutoMapperProfile.Mapper.Map<PSAzureOperationResponse>(op);
-                WriteObject(result);
+                if (NoWait.IsPresent)
+                {
+                    var op = this.VirtualMachineExtensionClient.BeginDeleteWithHttpMessagesAsync(
+                        this.ResourceGroupName,
+                        this.VMName,
+                        this.Name).GetAwaiter().GetResult();
+                    var result = ComputeAutoMapperProfile.Mapper.Map<PSAzureOperationResponse>(op);
+                    WriteObject(result);
+                }
+                else
+                {
+                    var op = this.VirtualMachineExtensionClient.DeleteWithHttpMessagesAsync(
+                        this.ResourceGroupName,
+                        this.VMName,
+                        this.Name).GetAwaiter().GetResult();
+                    var result = ComputeAutoMapperProfile.Mapper.Map<PSAzureOperationResponse>(op);
+                    WriteObject(result);
+                }
             });
         }
     }
